# Conflicts:
#	ladybug/pom.xml
#	ladybug/src/main/resources/springIbisTestTool.xml

+++ filesystem/src/main/java/org/frankframework/filesystem/smb/Samba1FileSystem.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2018 Nationale-Nederlanden, 2020-2022 WeAreFrank!
+   Copyright 2018 Nationale-Nederlanden, 2020-2024 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -21,6 +21,9 @@ import java.nio.file.DirectoryStream;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.Map;
+import java.util.NoSuchElementException;
+
+import jakarta.annotation.Nullable;
 
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -32,11 +35,17 @@ import jcifs.smb.SmbFileFilter;
 import jcifs.smb.SmbFileInputStream;
 import jcifs.smb.SmbFileOutputStream;
 import lombok.Getter;
+
 import org.frankframework.configuration.ConfigurationException;
-import org.frankframework.filesystem.FileSystemBase;
+import org.frankframework.filesystem.FileAlreadyExistsException;
+import org.frankframework.filesystem.FileNotFoundException;
+import org.frankframework.filesystem.AbstractFileSystem;
 import org.frankframework.filesystem.FileSystemException;
 import org.frankframework.filesystem.FileSystemUtils;
+import org.frankframework.filesystem.FolderAlreadyExistsException;
+import org.frankframework.filesystem.FolderNotFoundException;
 import org.frankframework.filesystem.IWritableFileSystem;
+import org.frankframework.filesystem.TypeFilter;
 import org.frankframework.stream.Message;
 import org.frankframework.util.CredentialFactory;
 
@@ -45,7 +54,7 @@ import org.frankframework.util.CredentialFactory;
  * <br/>
  * Only supports NTLM authentication.
  */
-public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritableFileSystem<SmbFile> {
+public class Samba1FileSystem extends AbstractFileSystem<SmbFile> implements IWritableFileSystem<SmbFile> {
 	private final @Getter String domain = "SMB";
 
 	private @Getter String share = null;
@@ -73,7 +82,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 		CredentialFactory cf = new CredentialFactory(getAuthAlias(), getUsername(), getPassword());
 		if (StringUtils.isNotEmpty(cf.getUsername())) {
 			auth = new NtlmPasswordAuthentication(getAuthenticationDomain(), cf.getUsername(), cf.getPassword());
-			log.debug("setting authentication to [" + auth.toString() + "]");
+			log.debug("setting authentication to [{}]", auth);
 		}
 	}
 
@@ -89,7 +98,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile toFile(String filename) throws FileSystemException {
+	public SmbFile toFile(@Nullable String filename) throws FileSystemException {
 		try {
 			return new SmbFile(smbContext, filename);
 		} catch (IOException e) {
@@ -98,14 +107,14 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile toFile(String folder, String filename) throws FileSystemException {
+	public SmbFile toFile(@Nullable String folder, @Nullable String filename) throws FileSystemException {
 		return toFile(folder+"/"+filename);
 	}
 
 	@Override
-	public DirectoryStream<SmbFile> listFiles(String folder) throws FileSystemException {
+	public DirectoryStream<SmbFile> list(String folder, TypeFilter filter) throws FileSystemException {
 		try {
-			return FileSystemUtils.getDirectoryStream(new SmbFileIterator(folder));
+			return FileSystemUtils.getDirectoryStream(new SmbFileIterator(folder, filter));
 		} catch (IOException e) {
 			throw new FileSystemException(e);
 		}
@@ -147,7 +156,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	public void deleteFile(SmbFile f) throws FileSystemException {
 		try {
 			if (!f.exists()) {
-				throw new FileSystemException("File ["+f.getName()+"] not found");
+				throw new FileNotFoundException("File ["+f.getName()+"] not found");
 			}
 			if (f.isFile()) {
 				f.delete();
@@ -156,11 +165,12 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 						"Trying to remove [" + f.getName() + "] which is a directory instead of a file");
 			}
 		} catch (SmbException e) {
-			throw new FileSystemException("Could not delete file [" + getCanonicalName(f) + "]: " + e.getMessage());
+			throw new FileSystemException("Could not delete file [" + getCanonicalNameOrErrorMessage(f) + "]: " + e.getMessage(), e);
 		}
 	}
 
-	private boolean isFolder(SmbFile f) throws FileSystemException {
+	@Override
+	public boolean isFolder(SmbFile f) throws FileSystemException {
 		try {
 			return f.isDirectory();
 		} catch (SmbException e) {
@@ -177,7 +187,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	public void createFolder(String folder) throws FileSystemException {
 		try {
 			if(folderExists(folder)) {
-				throw new FileSystemException("Create directory for [" + folder + "] has failed. Directory already exists.");
+				throw new FolderAlreadyExistsException("Create directory for [" + folder + "] has failed. Directory already exists.");
 			}
 			if (isForce) {
 				toFile(folder).mkdirs();
@@ -193,15 +203,14 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	public void removeFolder(String folder, boolean removeNonEmptyFolder) throws FileSystemException {
 		String normalized = FilenameUtils.normalizeNoEndSeparator(folder, true) + "/";
 		try {
-			if (folderExists(normalized)) {
-				if(!removeNonEmptyFolder && listFiles(folder).iterator().hasNext()) {
-					throw new FileSystemException("Cannot remove folder [" + folder + "]. Directory not empty.");
-				}
-
-				toFile(normalized).delete();
-			} else {
-				throw new FileSystemException("Cannot remove folder [" + normalized + "]. Directory does not exist.");
+			if (!folderExists(normalized)) {
+				throw new FolderNotFoundException("Cannot remove folder [" + normalized + "]. Directory does not exist.");
 			}
+			if(!removeNonEmptyFolder && list(folder, TypeFilter.FILES_ONLY).iterator().hasNext()) {
+				throw new FileSystemException("Cannot remove folder [" + folder + "]. Directory not empty.");
+			}
+
+			toFile(normalized).delete();
 		} catch (SmbException e) {
 			throw new FileSystemException(e);
 		}
@@ -209,6 +218,12 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 	@Override
 	public SmbFile renameFile(SmbFile source, SmbFile destination) throws FileSystemException {
+		if (!exists(source)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(source) + "], cannot rename.");
+		}
+		if (exists(destination)) {
+			throw new FileAlreadyExistsException("Cannot rename [" + getName(source) + "], destination [" + getName(destination) + "] already exists");
+		}
 		try {
 			source.renameTo(destination);
 			return destination;
@@ -218,7 +233,18 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile moveFile(SmbFile f, String destinationFolder, boolean createFolder, boolean resultantMustBeReturned) throws FileSystemException {
+	public SmbFile moveFile(SmbFile f, String destinationFolder, boolean createFolder) throws FileSystemException {
+		if (!exists(f)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(f) + "], cannot move.");
+		}
+		if (!folderExists(destinationFolder)) {
+			if(createFolder) {
+				createFolder(destinationFolder);
+			} else {
+				throw new FolderNotFoundException("destination does not exist");
+			}
+		}
+
 		SmbFile dest = toFile(destinationFolder, f.getName());
 		try {
 			f.renameTo(dest);
@@ -229,12 +255,15 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile copyFile(SmbFile f, String destinationFolder, boolean createFolder, boolean resultantMustBeReturned) throws FileSystemException {
+	public SmbFile copyFile(SmbFile f, String destinationFolder, boolean createFolder) throws FileSystemException {
+		if (!exists(f)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(f) + "], cannot copy.");
+		}
 		if (!folderExists(destinationFolder)) {
 			if(createFolder) {
 				createFolder(destinationFolder);
 			} else {
-				throw new FileSystemException("destination does not exist");
+				throw new FolderNotFoundException("destination does not exist");
 			}
 		}
 
@@ -255,23 +284,21 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 	private class SmbFileIterator implements Iterator<SmbFile> {
 
-		private SmbFile[] files;
+		private final SmbFile[] files;
 		private int i = 0;
 
-		public SmbFileIterator(String folder) throws IOException {
-			SmbFileFilter filter = new SmbFileFilter() {
-
-				@Override
-				public boolean accept(SmbFile file) throws SmbException {
-					return (!isListHiddenFiles() && !file.isHidden()) && file.isFile();
-				}
+		public SmbFileIterator(String folder, TypeFilter typeFilter) throws IOException {
+			SmbFileFilter filter = switch (typeFilter) {
+				case FILES_ONLY -> file -> (!isListHiddenFiles() && !file.isHidden()) && file.isFile();
+				case FOLDERS_ONLY -> file -> (!isListHiddenFiles() && !file.isHidden()) && file.isDirectory();
+				case FILES_AND_FOLDERS -> file -> (!isListHiddenFiles() && !file.isHidden());
 			};
 			SmbFile f = smbContext;
-			if(StringUtils.isNotBlank(folder)) {
-				String normalizedFolder = FilenameUtils.normalizeNoEndSeparator(folder, true)+"/";
+			if (StringUtils.isNotBlank(folder)) {
+				String normalizedFolder = FilenameUtils.normalizeNoEndSeparator(folder, true) + "/";
 				f = new SmbFile(smbContext, normalizedFolder);
 			}
-			this.files = f.listFiles(filter);
+			files = f.listFiles(filter);
 		}
 
 		@Override
@@ -281,6 +308,9 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 		@Override
 		public SmbFile next() {
+			if (i >= files.length) {
+				throw new NoSuchElementException();
+			}
 			return files[i++];
 		}
 
@@ -290,7 +320,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 			try {
 				deleteFile(file);
 			} catch (FileSystemException e) {
-				log.warn("unable to delete file ["+getCanonicalName(file)+"]", e);
+				log.warn("unable to delete file [{}]", getCanonicalNameOrErrorMessage(file), e);
 			}
 		}
 	}
@@ -306,6 +336,9 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 	@Override
 	public String getName(SmbFile f) {
+		if(f.getName().endsWith("/")) {
+			return StringUtils.chop(f.getName());
+		}
 		return f.getName();
 	}
 
@@ -325,6 +358,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
+	@Nullable
 	public Map<String, Object> getAdditionalFileProperties(SmbFile file) {
 		return null;
 	}
